requirement.txt
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
attrs
babel
beautifulsoup4
bleach
certifi
cffi
charset-normalizer
colorama
comm
contourpy
cycler
debugpy
decorator
defusedxml
executing
fastjsonschema
filelock
fonttools
fqdn
fsspec
h11
httpcore
httpx
idna
ipykernel
ipython
ipython_pygments_lexers
isoduration
jedi
Jinja2
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
kiwisolver
lark
MarkupSafe
matplotlib
matplotlib-inline
mistune
mpmath
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook_shim
numpy
packaging
pandas
pandocfilters
parso
pillow
platformdirs
prometheus_client
prompt_toolkit
psutil
pure_eval
pycparser
Pygments
pyparsing
python-dateutil
python-json-logger
pywinpty
PyYAML
pyzmq
referencing
requests
rfc3339-validator
rfc3986-validator
rfc3987-syntax
rpds-py
scikit-learn
scipy
seaborn
Send2Trash
setuptools
six
soupsieve
stack-data
sympy
terminado
threadpoolctl
tinycss2
torch
torchinfo
torchvision
tornado
tqdm
traitlets
typing_extensions
tzdata
uri-template
urllib3
wcwidth
webcolors
webencodings
websocket-client
